+++ fractal_server/app/runner/_local/_local_config.py
@@ -63,11 +63,10 @@ def get_local_backend_config(
     The sources for `parallel_tasks_per_job` attributes, starting from the
     highest-priority one, are
 
-    1. Properties in `wftask.meta`;
-    2. Properties in `wftask.task.meta` (this is already included in
-       overridden_meta);
-    3. The general content of the local-backend configuration file;
-    4. The default value (`None`).
+    1. Properties in `wftask.meta` (which, for `WorkflowTask`s added through
+       `Workflow.insert_task`, also includes `wftask.task.meta`);
+    2. The general content of the local-backend configuration file;
+    3. The default value (`None`).
 
     Arguments:
         wftask:
@@ -84,8 +83,8 @@ def get_local_backend_config(
     key = "parallel_tasks_per_job"
     default = None
 
-    if wftask.overridden_meta and key in wftask.overridden_meta.keys():
-        parallel_tasks_per_job = wftask.overridden_meta[key]
+    if wftask.meta and key in wftask.meta:
+        parallel_tasks_per_job = wftask.meta[key]
     else:
         if not config_path:
             settings = Inject(get_settings)
@@ -104,5 +103,4 @@ def get_local_backend_config(
                 )
 
             parallel_tasks_per_job = env.get(key, default)
-
     return LocalBackendConfig(parallel_tasks_per_job=parallel_tasks_per_job)